requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==3.0.4
Django==3.2.7
djangorestframework==3.12.4
idna==2.8
Pillow==8.3.2
postgres==4.0
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pytz==2021.3
requests==2.22.0
sqlparse==0.4.2
urllib3==1.25.11
